+++ src/log.js
@@ -27,7 +27,7 @@
 function constructor() {
   let log = require(`loglevel`).noConflict();
   let loglevelMessagePrefix = require(`@natlibfi/loglevel-message-prefix`);
-  log.setLevel(2);
+  log.setLevel(`info`);
 
   loglevelMessagePrefix(log, {
     prefixes: [`level`],
@@ -40,7 +40,7 @@ function constructor() {
       DEBUG: `debug`,
       INFO: `info`,
       WARN: `warn`,
-      ERROR: `error`
+      ERROR: `error`,
     }
   });